+++ src/app/core/auth/auth.service.ts
@@ -5,10 +5,11 @@ import { BehaviorSubject, Observable } from 'rxjs';
 
 import { User } from 'src/app/core/models/user';
 import { UserData } from 'src/app/core/data/user/user-data.interface';
-import { AuthResponseData } from 'src/app/core/http/api/api.interfaces';
+import { AuthResponseData, AuthTokenStructure } from 'src/app/core/auth/auth.interfaces';
 import { ApiService } from 'src/app/core/http/api/api.service';
 import { CookieDataService } from 'src/app/core/services/cookie/cookie-data.service';
 import { StaticService } from 'src/app/config/static.service';
+import jwt_decode from 'jwt-decode';
 
 /**
  * auth.service.ts
@@ -23,9 +24,6 @@ export class AuthService {
   public user$ = new BehaviorSubject<User>(null);
   // Userdata, streams with every change
   public userAuthenticated$: Observable<UserData>;
-  // TODO: userForInterceptor$ löschen, wenn userID im Token ist
-  // (wird nur noch im Interceptor verwendet, um die Anfrage auf die UserId zuzulassen)
-  public userForInterceptor$ = new BehaviorSubject<User>(null);
   private tokenExpirationTimer: any;
 
   lnkAfterLogout = this.staticConfig.getRoutingInfo().lnkAfterLogout;
@@ -46,50 +44,32 @@ export class AuthService {
     );
   }
 
-  login(email: string, password: string): Observable<any> {
+  login(email: string, password: string): Observable<User> {
     return this.apiService.loginUser(email, password).pipe(
-      concatMap((serverResponse: AuthResponseData) => {
-        console.table(serverResponse);
-        const expirationDate = new Date(
-          new Date().getTime() + +serverResponse.expires_in * 1000
-        );
+      map((serverResponse: AuthResponseData) => {
+        // console.table(serverResponse);
+        const expirationDate = new Date(new Date().getTime() + +serverResponse.expires_in * 1000);
+        const decoded = this.getDecodedToken(serverResponse.access_token);
+        //console.log('Decoded Token: ', decoded);
         const user = new User(
-          0,
+          decoded.user_id,
           email,
-          undefined,
+          decoded.user_name,
           serverResponse.access_token,
           expirationDate
         );
-
-        // Sonst kann das Token über den Interceptor nicht mitgesendet werden
-        this.userForInterceptor$.next(user); // Todo: Delete
-
-        // Todo: Die UserID und Namen im Token mitsenden,
-        // dann muss man hier keine weiteren Abfragen machen
-        return this.apiService.getUserByEmail(email).pipe(
-          tap((userDataArr) => {
-            //console.log('UserData from Server: ', userDataArr);
-            let userData = userDataArr[0];
-            if (userData) {
-              user.id = userData.id;
-              user.name = userData.name;
-              this.user$.next(user);
-              this.userForInterceptor$.next(user);
-              this.cookieDataService.setLocalStorageItem('userData',JSON.stringify(user) );
-              const expirationDuration =
-                new Date(user.tokenExpirationDate).getTime() -
-                new Date().getTime();
-              this.autoLogout(expirationDuration);
-            }
-          })
-        );
+        this.user$.next(user);
+        this.cookieDataService.setLocalStorageItem('userData', JSON.stringify(user));
+        const expirationDuration =
+          new Date(user.tokenExpirationDate).getTime() - new Date().getTime();
+        this.autoLogout(expirationDuration);
+        return user;
       })
     );
   }
 
   logout() {
     this.user$.next(null);
-    this.userForInterceptor$.next(null);
     this.cookieDataService.deleteLocalStorageItem('userData');
     if (this.tokenExpirationTimer) {
       clearTimeout(this.tokenExpirationTimer);
@@ -116,7 +96,6 @@ export class AuthService {
 
     if (!userData || !userData._token) {
       this.user$.next(null);
-      this.userForInterceptor$.next(null);
       return;
     }
 
@@ -132,6 +111,46 @@ export class AuthService {
     );
 
     this.user$.next(user);
-    this.userForInterceptor$.next(user);
   }
+
+  private getDecodedToken(token: string): AuthTokenStructure {
+    try {
+      return jwt_decode(token);
+    } catch (Error) {
+      return null;
+    }
+  }
+
+  // TO_DELETE
+  // login_alt(email: string, password: string): Observable<any> {
+  //   return this.apiService.loginUser(email, password).pipe(
+  //     concatMap((serverResponse: AuthResponseData) => {
+  //       console.table(serverResponse);
+  //       const expirationDate = new Date(new Date().getTime() + +serverResponse.expires_in * 1000);
+  //       const user = new User(0, email, undefined, serverResponse.access_token, expirationDate);
+
+  //       // Sonst kann das Token über den Interceptor nicht mitgesendet werden
+  //        this.userForInterceptor$.next(user); // Todo: Delete
+
+  //       // Todo: Die UserID und Namen im Token mitsenden,
+  //       // dann muss man hier keine weiteren Abfragen machen
+  //       return this.apiService.getUserByEmail(email).pipe(
+  //         tap((userDataArr) => {
+  //           //console.log('UserData from Server: ', userDataArr);
+  //           let userData = userDataArr[0];
+  //           if (userData) {
+  //             user.id = userData.id;
+  //             user.name = userData.name;
+  //             this.user$.next(user);
+  //              this.userForInterceptor$.next(user); //Delete
+  //             this.cookieDataService.setLocalStorageItem('userData', JSON.stringify(user));
+  //             const expirationDuration =
+  //               new Date(user.tokenExpirationDate).getTime() - new Date().getTime();
+  //             this.autoLogout(expirationDuration);
+  //           }
+  //         })
+  //       );
+  //     })
+  //   );
+  // }
 }